requirements.txt
Django>=4.2,<5.0
django-tenants>=3.5
djangorestframework>=3.14
dj-rest-auth>=5.0
django-allauth>=0.58
workos>=5.0
psycopg2-binary>=2.9
gunicorn>=21.2
dj-database-url>=2.1
python-dotenv>=1.0
google-auth>=2.29
google-api-python-client>=2.120
google-cloud-dns>=0.34
google-cloud-tasks>=2.16
